Make Tracer a class based decorator
